Change one/zero size & threshold between the two
This adds a small black border of 2px to the right and bottom of the
image, but I think it looks much better with the slightly smaller
"pixels". Also changing the threshold before a 1 becomes a 0 to add some
more 1's to the image, which I think makes the resulting image a little
sharper in appearance.